+++ mobile/android/android-components/components/browser/engine-gecko-nightly/src/main/java/mozilla/components/browser/engine/gecko/GeckoEngineView.kt
@@ -91,6 +91,11 @@ class GeckoEngineView @JvmOverloads constructor(
         currentSession?.apply { unregister(observer) }
     }
 
+    override fun canScrollVerticallyUp(): Boolean {
+        val result = currentSession?.let { it.scrollY > 0 } != false
+        return result
+    }
+
     override fun canScrollVerticallyDown() = true // waiting for this issue https://bugzilla.mozilla.org/show_bug.cgi?id=1507569
 
     override fun setVerticalClipping(clippingHeight: Int) {