+++ mobile/android/android-components/components/browser/engine-system/src/main/java/mozilla/components/browser/engine/system/SystemEngineView.kt
@@ -640,6 +640,8 @@ class SystemEngineView @JvmOverloads constructor(
         // no-op
     }
 
+    override fun canScrollVerticallyUp() = session?.webView?.canScrollVertically(-1) ?: false
+
     override fun canScrollVerticallyDown() = session?.webView?.canScrollVertically(1) ?: false
 
     @Suppress("Deprecation")